ml/requirements.txt
# ─── Core ML ────────────────────────────────────────────────────────────────
torch
torchvision
torchaudio

# ─── YOLO (pneumonia detection) ─────────────────────────────────────────────
ultralytics==8.3.40

# ─── Medical imaging (3D MRI segmentation) ──────────────────────────────────
monai==1.3.2
nibabel==5.3.1
SimpleITK==2.4.0
pydicom==3.0.1

# ─── Model architectures (ViT, EfficientNet) ────────────────────────────────
timm==1.0.11

# ─── Augmentations & image processing ───────────────────────────────────────
albumentations==1.4.21
opencv-python-headless==4.10.0.84
Pillow==11.0.0

# ─── Data & numerics ─────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1

# ─── Visualisation ───────────────────────────────────────────────────────────
matplotlib==3.9.3
seaborn==0.13.2

# ─── Experiment tracking ─────────────────────────────────────────────────────
clearml==1.16.4

# ─── Export ──────────────────────────────────────────────────────────────────
onnx==1.17.0
onnxruntime

# ─── Dataset download ────────────────────────────────────────────────────────
roboflow
python-dotenv

# ─── Utilities ───────────────────────────────────────────────────────────────
tqdm==4.67.1
rich==13.9.4
PyYAML==6.0.2
